REFACTOR: refactoring navigation
