+++ src/modules/App/index.tsx
@@ -1,7 +1,7 @@
 import 'react-native-gesture-handler';
 import React from 'react';
 
-import TabNavigation from 'navigation/AppNavigation';
+import TabNavigation from 'navigation/TabNavigation';
 
 const App: React.FC = () => {
   return <TabNavigation />;
