+++ src/navigation/StackNavigators/index.tsx
@@ -1,42 +1,38 @@
 import 'react-native-gesture-handler';
 import React from 'react';
 import { Platform } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
-import Ionicons from 'react-native-vector-icons/Ionicons';
 import { format } from 'date-fns';
 
 import THEME from 'theme';
+import BookedScreen from 'screens/BookedScreen';
+import PostScreen from 'screens/PostScreen';
 import MainScreen from 'screens/MainScreen';
 import AboutScreen from 'screens/AboutScreen';
-import BookedScreen from 'screens/BookedScreen';
 import CreateScreen from 'screens/CreateScreen';
-import PostScreen from 'screens/PostScreen';
 import HeaderNavButtons from 'components/HeaderNavButtons';
+
+const SCREEN_OPTIONS = {
+  headerStyle: {
+    backgroundColor:
+      Platform.OS === 'android' ? THEME.MAIN_COLOR : THEME.BORDER_COLOR,
+  },
+  headerTintColor:
+    Platform.OS === 'android' ? THEME.WHITE_COLOR : THEME.TEXT_COLOR,
+  headerTitleStyle: {
+    fontSize: 24,
+    fontFamily: 'BadScript-Regular',
+  },
+};
 
 const HomeStack = createStackNavigator();
 const BookedStack = createStackNavigator();
-const IosTab = createBottomTabNavigator();
-const AndroidTab = createMaterialBottomTabNavigator();
 
-const HomeStackScreen: React.FC = () => {
+export const HomeStackScreen: React.FC = () => {
   return (
     <HomeStack.Navigator
       initialRouteName="MainScreen"
-      screenOptions={{
-        headerStyle: {
-          backgroundColor:
-            Platform.OS === 'android' ? THEME.MAIN_COLOR : THEME.BORDER_COLOR,
-        },
-        headerTintColor:
-          Platform.OS === 'android' ? THEME.WHITE_COLOR : THEME.TEXT_COLOR,
-        headerTitleStyle: {
-          fontSize: 24,
-          fontFamily: 'BadScript-Regular',
-        },
-      }}>
+      screenOptions={SCREEN_OPTIONS}>
       <HomeStack.Screen
         name="MainScreen"
         component={MainScreen}
@@ -122,21 +118,9 @@ const HomeStackScreen: React.FC = () => {
   );
 };
 
-const BookedStackScreen: React.FC = () => {
+export const BookedStackScreen: React.FC = () => {
   return (
-    <BookedStack.Navigator
-      screenOptions={{
-        headerStyle: {
-          backgroundColor:
-            Platform.OS === 'android' ? THEME.MAIN_COLOR : THEME.BORDER_COLOR,
-        },
-        headerTintColor:
-          Platform.OS === 'android' ? THEME.WHITE_COLOR : THEME.TEXT_COLOR,
-        headerTitleStyle: {
-          fontSize: 24,
-          fontFamily: 'BadScript-Regular',
-        },
-      }}>
+    <BookedStack.Navigator screenOptions={SCREEN_OPTIONS}>
       <BookedStack.Screen
         name="BookedScreen"
         component={BookedScreen}
@@ -151,7 +135,7 @@ const BookedStackScreen: React.FC = () => {
           ),
         }}
       />
-      <HomeStack.Screen
+      <BookedStack.Screen
         name="PostScreen"
         component={PostScreen}
         options={({ route }: any) => ({
@@ -171,73 +155,3 @@ const BookedStackScreen: React.FC = () => {
     </BookedStack.Navigator>
   );
 };
-
-const AndroidTabNavigation: React.FC = () => {
-  return (
-    <AndroidTab.Navigator
-      initialRouteName="Home"
-      activeColor={THEME.WHITE_COLOR}
-      inactiveColor={THEME.MAIN_COLOR4}
-      barStyle={{ backgroundColor: THEME.MAIN_COLOR }}>
-      <AndroidTab.Screen
-        name="Home"
-        component={HomeStackScreen}
-        options={{
-          tabBarLabel: 'Home',
-          tabBarIcon: ({ color }) => (
-            <Ionicons name="home" color={color} size={24} />
-          ),
-        }}
-      />
-      <AndroidTab.Screen
-        name="Bookmarked"
-        component={BookedStackScreen}
-        options={{
-          tabBarLabel: 'Bookmarked',
-          tabBarIcon: ({ color }) => (
-            <Ionicons name="star" color={color} size={24} />
-          ),
-        }}
-      />
-    </AndroidTab.Navigator>
-  );
-};
-
-const IosTabNavigation: React.FC = () => {
-  return (
-    <IosTab.Navigator
-      screenOptions={({ route }) => ({
-        tabBarIcon: ({ color, size }) => {
-          let iconName = '';
-
-          if (route.name === 'Home') {
-            iconName = 'home';
-          } else if (route.name === 'Bookmarked') {
-            iconName = 'star';
-          }
-          return <Ionicons name={iconName} size={size} color={color} />;
-        },
-      })}
-      tabBarOptions={{
-        activeTintColor: THEME.MAIN_COLOR,
-        inactiveTintColor: THEME.TEXT_COLOR,
-      }}>
-      <IosTab.Screen name="Home" component={HomeStackScreen} />
-      <IosTab.Screen name="Bookmarked" component={BookedStackScreen} />
-    </IosTab.Navigator>
-  );
-};
-
-const TabNavigation: React.FC = () => {
-  return (
-    <NavigationContainer>
-      {Platform.OS === 'android' ? (
-        <AndroidTabNavigation />
-      ) : (
-        <IosTabNavigation />
-      )}
-    </NavigationContainer>
-  );
-};
-
-export default TabNavigation;
